simplifying defaults (1/2)
